graphql_server/requirements.txt
fastapi
strawberry-graphql
requests
uvicorn